factoring execute entry point custom after server HTTP running
