+++ lib/build-files/content-files/_index.js
@@ -120,37 +120,31 @@ const startHTTP = function() {
 
 		if( typeof process.env.ENTRY_POINT === "string" ) {
 
-			const pathEntryPoint = pathResolver.join(
-				__dirname , process.env.ENTRY_POINT.trim()
-			) ;
+			const isFound = require('./mvc-vendor/execute-custom-entry-point')({
+				pathFile: process.env.ENTRY_POINT.trim(),
+				httpListener,
+				server,
+				router
+			}) ;
 
-			if( fs.existsSync( pathEntryPoint ) ) {
-
-				let outputEntryPoint = require( pathEntryPoint ) ;
-
-				while( outputEntryPoint instanceof Function ) {
-
-					outputEntryPoint = outputEntryPoint(
-						httpListener, // http.Server
-						server, // net.Server
-						router // express.Router
-					) ;
-				}
-			} else {
+			if( !isFound ) {
 
 				const message = `error: entry point file define from .env not found with: ${process.env.ENTRY_POINT}` ;
 
 				if( CMP.isDev ) {
 
+
 					console.log( colorLog.red.bold( message ) ) ;
 
 				} else {
 
-					console.log( colorLog.red.bold( message ) ) ;
+					console.log( message ) ;
 				}
 
 				process.exit() ;
+
 			}
+
 		}
 
 	} ) ;
